bug fixes for project1

+++ project1/RightTriangle.java
@@ -1,9 +1,9 @@
 public class RightTriangle {
 	public static void main(String[] args) {
 
-		int side0 = Long.parseLong(args[0]);
-		int side1 = Long.parseLong(args[1]);
-		int side2 = Long.parseLong(args[2]);
+		long side0 = Long.parseLong(args[0]);
+		long side1 = Long.parseLong(args[1]);
+		long side2 = Long.parseLong(args[2]);
 
 		boolean check1 = (side0 * side0) == (side1 * side1) + (side2 * side2);
 		boolean check2 = (side1 * side1) == (side0 * side0) + (side2 * side2);
